[PW-6705] - Refactor AdyenMotoConfigProvider.php
