+++ Gateway/Request/ShopperInteractionDataBuilder.php
@@ -11,8 +11,8 @@
 
 namespace Adyen\Payment\Gateway\Request;
 
+use Adyen\Payment\Model\Ui\Adminhtml\AdyenMotoConfigProvider;
 use Adyen\Payment\Model\Ui\AdyenCcConfigProvider;
-use Adyen\Payment\Model\Ui\AdyenMotoConfigProvider;
 use Adyen\Payment\Model\Ui\AdyenOneclickConfigProvider;
 use Adyen\Payment\Model\Ui\AdyenPayByLinkConfigProvider;
 use Magento\Framework\Exception\LocalizedException;
@@ -22,6 +22,10 @@ use Magento\Payment\Gateway\Request\BuilderInterface;
 
 class ShopperInteractionDataBuilder implements BuilderInterface
 {
+    const SHOPPER_INTERACTION_MOTO = 'Moto';
+    const SHOPPER_INTERACTION_CONTAUTH = 'ContAuth';
+    const SHOPPER_INTERACTION_ECOMMERCE = 'Ecommerce';
+
     /**
      * @var \Magento\Framework\App\State
      */
@@ -56,16 +60,16 @@ class ShopperInteractionDataBuilder implements BuilderInterface
         }
 
         // Ecommerce is the default shopperInteraction
-        $shopperInteraction = 'Ecommerce';
+        $shopperInteraction = self::SHOPPER_INTERACTION_ECOMMERCE;
 
         if ($paymentMethod == AdyenMotoConfigProvider::CODE &&
             $this->appState->getAreaCode() == \Magento\Framework\App\Area::AREA_ADMINHTML) {
             // Backend CC orders are MOTO
-            $shopperInteraction = "Moto";
+            $shopperInteraction = self::SHOPPER_INTERACTION_MOTO;
         } elseif ($paymentMethod == AdyenOneclickConfigProvider::CODE
             || $paymentMethod == AdyenCcConfigProvider::CC_VAULT_CODE) {
             // OneClick and Vault are ContAuth
-            $shopperInteraction = 'ContAuth';
+            $shopperInteraction = self::SHOPPER_INTERACTION_CONTAUTH;
         }
 
         $requestBody['body']['shopperInteraction'] = $shopperInteraction;
